Implemented a KeyboardInput class with callback listeners similar to MouseInput. Not all KEYS have been registered yet and will follow in the next few updates. Also started with the official versioning. Starting from this iteration forward it will follow the MAJOR.MINOR.MICRO pattern

+++ engine/src/main/java/nl/arkenbout/geoffrey/angel/engine/core/input/MouseListener.java
@@ -9,7 +9,7 @@ public interface MouseListener {
     void onRightButtonDown(Vector2f mouseDelta);
     void onRightUp();
 
-    void onScroll(double scrollDelta);
+    void onScroll(float scrollDelta);
 
     void onMouseMove(Vector2f mouseDelta);
 }
